Use int for number of octaves. Don't why the original author chose to use float here..

+++ libgg/Noise.h
@@ -1,20 +1,20 @@
 #pragma once
 
 float octave_noise_1d(
-	const float octaves,
+	const int octaves,
 	const float persistence,
 	const float scale,
 	const float x);
 
 float octave_noise_2d(
-	const float octaves,
+	const int octaves,
 	const float persistence,
 	const float scale,
 	const float x,
 	const float y);
 
 float octave_noise_3d(
-	const float octaves,
+	const int octaves,
 	const float persistence,
 	const float scale,
 	const float x,
@@ -22,7 +22,7 @@ float octave_noise_3d(
 	const float z);
 
 float octave_noise_4d(
-	const float octaves,
+	const int octaves,
 	const float persistence,
 	const float scale,
 	const float x,
@@ -33,7 +33,7 @@ float octave_noise_4d(
 //
 
 float scaled_octave_noise_1d(
-	const float octaves,
+	const int octaves,
 	const float persistence,
 	const float scale,
 	const float loBound,
@@ -41,7 +41,7 @@ float scaled_octave_noise_1d(
 	const float x);
 
 float scaled_octave_noise_2d(
-	const float octaves,
+	const int octaves,
 	const float persistence,
 	const float scale,
 	const float loBound,
@@ -50,7 +50,7 @@ float scaled_octave_noise_2d(
 	const float y);
 
 float scaled_octave_noise_3d(
-	const float octaves,
+	const int octaves,
 	const float persistence,
 	const float scale,
 	const float loBound,
@@ -60,7 +60,7 @@ float scaled_octave_noise_3d(
 	const float z);
 
 float scaled_octave_noise_4d(
-	const float octaves,
+	const int octaves,
 	const float persistence,
 	const float scale,
 	const float loBound,
